university_admission_vs_code/requirements.txt
flask
flask-login
flask-sqlalchemy
gunicorn
pillow
werkzeug
email-validator